extracted creation of kwargs for builder function
