+++ nicegui/page_args.py
@@ -1,7 +1,8 @@
 from __future__ import annotations
 
+import inspect
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING, Any, Callable, Dict
 
 if TYPE_CHECKING:
     from .elements.sub_pages import SubPages
@@ -9,6 +10,21 @@ if TYPE_CHECKING:
 from starlette.datastructures import QueryParams
 
 from .dataclasses import KWONLY_SLOTS
+
+
+@dataclass(**KWONLY_SLOTS)
+class RouteMatch:
+    """Information about a matched route."""
+    path: str
+    '''The sub-path that actually matched (e.g., "/user/123")'''
+    pattern: str
+    '''The original route pattern (e.g., "/user/{id}")'''
+    builder: Callable
+    '''The function to call to build the page'''
+    parameters: Dict[str, str]
+    '''The extracted parameters (name -> value) from the path (e.g., ``{"id": "123"}``)'''
+    query_params: QueryParams
+    '''The query parameters from the URL'''
 
 
 @dataclass(**KWONLY_SLOTS)
@@ -28,3 +44,60 @@ class PageArgs:
     '''Query parameters from the request URL.'''
     data: dict[str, Any]
     '''Arbitrary data passed to the ui.sub_pages element.'''
+
+    @classmethod
+    def _from_route_match(cls, route_match: RouteMatch, frame: SubPages, data: dict[str, Any]) -> PageArgs:
+        """Create a PageArgs instance from a RouteMatch and SubPages frame.
+
+        :param route_match: The RouteMatch containing path info and parameters
+        :param frame: The SubPages instance currently executing this page
+        :param data: Arbitrary data passed to the sub_pages element
+        :return: A new PageArgs instance
+        """
+        return cls(
+            path=route_match.path,
+            frame=frame,
+            path_parameters=route_match.parameters or {},
+            query_parameters=route_match.query_params,
+            data=data
+        )
+
+    @staticmethod
+    def _convert_parameter(value: str, param_type: type) -> Any:
+        """Convert a string parameter to the specified type (``str``, ``int``, or ``float``).
+
+        :param value: the string value to convert
+        :param param_type: the type to convert to
+        :return: the converted value
+        """
+        if param_type is str or param_type is inspect.Parameter.empty:
+            return value
+        elif param_type is int:
+            return int(value)
+        elif param_type is float:
+            return float(value)
+        return value
+
+    @classmethod
+    def build_kwargs(cls, route_match: RouteMatch, frame: SubPages, data: dict[str, Any]) -> dict[str, Any]:
+        """Build keyword arguments for the route builder function.
+
+        :param route_match: The RouteMatch containing path info and parameters
+        :param frame: The SubPages instance currently executing this page
+        :param data: Arbitrary data passed to the sub_pages element
+        :return: Dictionary of keyword arguments for the builder function
+        """
+        parameters = inspect.signature(route_match.builder).parameters
+        kwargs = {}
+
+        for name, param in parameters.items():
+            if param.annotation is cls:
+                kwargs[name] = cls._from_route_match(route_match, frame, data)
+            elif name in data:
+                kwargs[name] = data[name]
+            elif route_match.parameters and name in route_match.parameters:
+                kwargs[name] = cls._convert_parameter(route_match.parameters[name], param.annotation)
+            elif name in route_match.query_params:
+                kwargs[name] = cls._convert_parameter(route_match.query_params[name], param.annotation)
+
+        return kwargs
